front: standardize utils file naming and remove the unused function
