+++ front/tests/003-study-management.spec.ts
@@ -5,8 +5,8 @@ import type { Project, Study } from 'common/api/osrdEditoastApi';
 import test from './logging-fixture';
 import StudyPage from './pages/operational-studies/study-page';
 import { generateUniqueName, getTranslations } from './utils';
-import { getProject } from './utils/api-setup';
-import { formatDateToDayMonthYear } from './utils/date';
+import { getProject } from './utils/api-utils';
+import { formatDateToDayMonthYear } from './utils/date-utils';
 import readJsonFile from './utils/file-utils';
 import { createStudy } from './utils/setup-utils';
 import { deleteStudy } from './utils/teardown-utils';
