+++ front/tests/010-op-route-tab.spec.ts
@@ -6,7 +6,7 @@ import OperationalStudiesPage from './pages/operational-studies/operational-stud
 import RouteTab from './pages/operational-studies/route-tab';
 import RollingStockSelector from './pages/rolling-stock/rolling-stock-selector';
 import { waitForInfraStateToBeCached } from './utils';
-import { getInfra } from './utils/api-setup';
+import { getInfra } from './utils/api-utils';
 import createScenario from './utils/scenario';
 import { deleteScenario } from './utils/teardown-utils';
